+++ app.py
@@ -12,13 +12,11 @@ app = Flask(__name__)
 
 Db = DBClient(host='localhost', port=5432,
               user='RestaurantAdmin', password='RestaurantAdmin',
-              db='OpenRestaurant', app=app)
+              db='OpenRestaurant')
 
 # models
-
 user = User()
 address = Address()
-
 
 
 # Route for seeing a data
